+++ src/com/hqyj/pojo/Feerecord.java
@@ -6,7 +6,7 @@ public class Feerecord {
     private String cno;
     private Integer beginid;
     private Integer endid;
-    private Integer price;
+    private Double price;
     private Integer state; // 0»ò1
     private String backup;
     private Date createtime;
@@ -21,8 +21,8 @@ public class Feerecord {
     public void setBeginid(Integer beginid) { this.beginid = beginid; }
     public Integer getEndid() { return endid; }
     public void setEndid(Integer endid) { this.endid = endid; }
-    public Integer getPrice() { return price; }
-    public void setPrice(Integer price) { this.price = price; }
+    public Double getPrice() { return price; }
+    public void setPrice(Double price) { this.price = price; }
     public Integer getState() { return state; }
     public void setState(Integer state) { this.state = state; }
     public String getBackup() { return backup; }
